replace particle filter by particle for now
